Revert "Plan 9 from Bell Labs Fourth Edition 2013-05-01"
This reverts commit e94a8e9b44de11e62dfd4566245554f8bf8e0295.

+++ arm/include/u.h
@@ -9,7 +9,7 @@ typedef	long long	vlong;
 typedef	unsigned long long uvlong;
 typedef unsigned long	uintptr;
 typedef unsigned long	usize;
-typedef	uint		Rune;
+typedef	ushort		Rune;
 typedef 	union FPdbleword FPdbleword;
 typedef long	jmp_buf[2];
 #define	JMPBUFSP	0
